Factor out the Named trait
