+++ lang/ast/src/traits/mod.rs
@@ -1,11 +1,13 @@
 mod has_span;
 mod has_type;
+pub mod named;
 mod occurs;
 mod shift;
 pub mod subst;
 
 pub use has_span::*;
 pub use has_type::*;
+pub use named::*;
 pub use occurs::*;
 pub use shift::*;
 pub use subst::*;
